ycombinator_click_prediction/requirements.txt
dj-database-url==0.4.0
Django==1.9.2
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
SQLAlchemy==1.0.11
pandas==0.17.1
lxml==3.5.0
html5lib
BeautifulSoup4
requests
django-registration-redux==1.1
django-crispy-forms==1.6.0
ipython
pyyaml
user-agents==1.0.1
django-user-agents==0.3.0
ua-parser==0.7.0
pycallgraph
graphviz
simserver
geocoder
dill
neo4j-driver